feat: add web player html and css to fix overscan
This is a very simple fix for display overscan that simply renders the
movie or TV show you want to watch in a styled HTML `<video>` tag that
adds padding to compensate for overscan. To adjust this padding, update
`styles.css`. To change the source of the `<video>`, update
`index.html`. Note that you'll probably want to open this HTML file in
your browser's full-screen mode (and hide the browser search bar ofc).